### 2022-11-15 #4933 @editorialbot
[PRE REVIEW]: GeoHexViz: A Python package for visualizing hexagonally binned geospatial data
**Submitting author:** @mrempel (<a href="http://orcid.org/0000-0002-6248-1722">Mark Rempel</a>)
**Repository:** https://github.com/mrempel/geohexviz
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.0
