### 2022-11-22 #4933 @lwasser
should pyopensci review this submission and then fast track to joss Martin?
This seems like it would be in scope for pyopensci.

On Tue, Nov 22, 2022, 02:15 Martin Fleischmann ***@***.***>
wrote:

> Hi @lwasser <https://github.com/lwasser>, @simonrp84
> <https://github.com/simonrp84>, @gassmoeller
> <https://github.com/gassmoeller>, would you be willing too review this
> submission for JOSS?
>
