### 2022-11-22 #4933 @mrempel
Hello @lwasser @danielskatz - The software (and associated paper) have not been submitted to pyopensci. My co-author and I would prefer to continue with the submission to JOSS at this time.
