### 2023-01-13 #5073 @editorialbot
[REVIEW]: GeoHexViz: A Python package for visualizing hexagonally binned geospatial data
**Submitting author:** @mrempel (<a href="http://orcid.org/0000-0002-6248-1722">Mark Rempel</a>)
**Repository:** https://github.com/mrempel/geohexviz
**Branch with paper.md** (empty if default branch):
**Version:** v1.0.2
