### 2023-01-24 mrempel/geohexviz#2 @gassmoeller
Conflicting dependencies for shapely in installation between environment.yml and setup.py
Part of openjournals/joss-reviews#5073.

When installing geohexviz environment.yml using conda it installs shapely 1.8.5 (as an implicit dependency of one of the other packages), while `pip setup.py` then replaces this version with shapely 1.7.1 (as specified in setup.py). However, shapely 1.7.1 seems to be incompatible with newer versions of geopandas. I see the error discussed here when running the forest fire example: https://stackoverflow.com/questions/73347010/why-do-i-get-an-error-when-trying-to-read-a-file-in-geopandas-included-datasets. When manually replacing 1.7.1 with 1.8.5 again I can run the example as expected.

I used the following workflow to install geohexviz:
`conda env create -f environment.yml -n geohexviz`
`conda activate geohexviz`
`conda install pip` (to make sure pip installs into the conda environment)
`pip install .` (inside geohexviz)

Was this a wrong workflow, or can you update the version of shapely specified in setup.py?
