### 2023-01-24 mrempel/geohexviz#3 @gassmoeller
Examples require changes to file names before they run successfully
Related to https://github.com/openjournals/joss-reviews/issues/5073

The paths to data files and output files in the examples are currently placeholder values. I understand that you want to point out the places to users where they have to adjust things for their application. However, in the current form that means that the first thing new users see of your software is an error message (not the great figures that the examples produce). I could see a place for template values in the jupyter notebooks (where it is immediately clear what needs to be changed, but especially in the JSON and python scripts this choice seems not intuitive. Would it make sense to change them to the (existing) data files and output files in the respective folders, so that the example results are easier to reproduce? You could still leave a comment that indicates these are the places to change for user applications.
