### 2023-01-24 #5073 @gassmoeller
@mrempel and @tony-zeidan: I went through the review checklist and wanted to congratulate you for this nicely written and documented piece of software. I think it already fulfills almost all the criteria for publication by JOSS and if you address the minor issues I raised in the repository the publication should be good to go. On a side note I think the main challenge for you going forward will be to keep it compatible to the numerous packages you depend on, I already see a number of deprecation warnings when executing the examples (let me know if you want me to post them as an issue). Anyway I hope you can keep the project alive and I may actually make use of it for some of my own work, although my community usually relies more on continuous interpolated maps.
