### 2023-01-24 #5073 @mrempel
@gassmoeller - In regards to your question about contribution and authorship, @tony-zeidan was my co-op student between Feb 2021 and Dec 2021. It is correct that Tony implemented the majority of the code within GeoHexViz; I conceived the software and its design requirements (based on previous research I published with colleagues, see Figure 8 in https://cradpdf.drdc-rddc.gc.ca/PDFS/unc356/p812844_A1b.pdf), defined the functional specification, selected and crafted the examples in the repository, helped Tony resolve issues during implementation, and drafted the article for this submission. As my student, it was Tony's responsibility to implement the code and manage the git repository (I provided Tony the manuscript to include in the repository) to provide him with experience in doing so. Upon completion of his co-op term, Tony transferred the repository to me to complete the submission to this journal.

I hope this helps - if you need further information, please do not hesitate to let me know.
