### 2023-01-25 #5073 @tony-zeidan
@gassmoeller Hi there! I can confirm what @mrempel stated in his reply to your question is indeed correct.

As to the deprecation warnings I'm sad to hear that but I'm pretty sure most of them are due to iterating over objects in a way that won't be supported going forward. These should be easy to fix. I could use a log of depreciation warnings to get an idea...

I look forward to seeing your GeoHexViz visualizations!
