### 2023-01-25 b1f558c
Revised based on @gassmoeller comments.
