+++ docs/paper/paper.md
@@ -43,7 +43,8 @@ Common across these applications is the requirement to have knowledge of geospat
 
 # Features
 
-In order to generate a publication-quality geospatial visualization, GeoHexViz requires an analyst to specify a set of *layers* &mdash; where each layer is defined as a "[group] of point, line, or area (polygon) features representing a particular class or type of real-world entities" [@layer-definition] &mdash; to be visualized. At a minimum, an analyst must specify one layer, the *hexbin layer*, through a set of required properties: first, a reference, such as the location of a CSV file, to the point-like data to be hexagonally binned; and second, references to the data containing the latitude, longitude, and value at each coordinate, e.g., columns names in the CSV file. If a value at each coordinate is not specified, a value of one is assumed by default (this seems a curious decision. A value of one means there is something there (forest fire, ...). wouldnt it make more sense to assume a default of 0? Not so relevant for this paper, but maybe a consideration for some future version of the software? Or is this meant for lists of events without a value column (i.e. a list of coordinates is interpreted as 'one event per coordinate')? This would make sense then). In addition, optional properties may be defined, such as the function to be applied to the values, i.e., `count`, `sum`, `max`, and the resolution of the hexagons (as defined by Uber H3). Default values for these optional properties are provided by GeoHexViz; for example, the default function applied is `sum`.
+In order to generate a publication-quality geospatial visualization, GeoHexViz requires an analyst to specify a set of *layers* &mdash; where each layer is defined as a "[group] of point, line, or area (polygon) features representing a particular class or type of real-world entities" [@layer-definition] &mdash; to be visualized. At a minimum, an analyst must specify one layer, the *hexbin layer*, through a set of required properties: first, a reference, such as the location of a CSV file, to the point-like data to be hexagonally binned; and second, references to the data containing the latitude, longitude, and value at each coordinate, e.g., columns names in the CSV file. If a value at each coordinate is not specified, a value of one is assumed by default, i.e., it is assumed there is one event per coordinate.
+In addition, optional properties may be defined, such as the function to be applied to the values, i.e., `count`, `sum`, `max`, and the resolution of the hexagons (as defined by Uber H3). Default values for these optional properties are provided by GeoHexViz; for example, the default function applied is `sum`.
 
 With this single layer, GeoHexViz generates a publication-quality visualization. However, if the visualization is not satisfactory, GeoHexViz enables an analyst to modify the visualization's properties. These properties may be categorized into two groups: first, those that use functionality provided by GeoHexViz that both integrates and extends its underlying libraries; and second, those that are passed directly to Plotly without modification. The first group of properties are: 
 
@@ -67,7 +68,7 @@ Allied aerial bombing in World War 2 occurred across a vast geographic region, w
 
 # Limitations
 
-GeoHexViz uses the GeoJSON format to plot data sets. With GeoJSON come difficulties when geometries cross the 180th meridian; they may be interpreted as wrapping around the globe [@macwright2016a]. In GeoHexViz, hexagonal geometries are supplied via Uber H3 [@uber-H3], and as such this issue has been discussed with its developers [@uberh3-validhex-issue]. GeoHexViz provides a simple solution to address this problem; it tracks geometries that cross the meridian and shifts their coordinates making all of the coordinates either positive or negative as previously proposed [@macwright2016a]. However, it should be noted that when hexagons contain either the North or South Pole, the 180th meridian issue persists, resulting in what appears to be a colour bleeding throughout the visualization and leaving a hexagon (or hexagons) empty. 
[…]
