+++ README.rst
@@ -353,7 +353,7 @@ by the Minister of National Defence, 2022.
     :target: https://github.com/mrempel/geohexviz/actions/workflows/geohexviz-automated-tests-conda.yml
 .. |cipip| image:: https://github.com/mrempel/geohexviz/actions/workflows/geohexviz-automated-tests-pip.yml/badge.svg
     :target: https://github.com/mrempel/geohexviz/actions/workflows/geohexviz-automated-tests-pip.yml
-.. |pythonver| image:: https://img.shields.io/badge/python-3.7,_3.8,_3.9-blue.svg
+.. |pythonver| image:: https://img.shields.io/badge/python-3.7,_3.8-blue.svg
 .. |geohexvizver| image:: https://img.shields.io/badge/geohexviz-v1.0.0-blue.svg
 .. |license| image:: https://img.shields.io/badge/License-BSD%203.0-blue.svg
     :target: https://github.com/mrempel/geohexviz/blob/master/LICENSE
