### 2023-01-26 6c78b6a
Big update
- Versioning should be fixed.
- Automated testing reflects version choices.
- As many deprecated/future warnings taken care of.
