+++ setup.py
@@ -1,4 +1,3 @@
-import pathlib
 from setuptools import setup, find_packages
 
 with open("README.rst") as readme_file:
