+++ requirements.txt
@@ -1,9 +1,14 @@
-setuptools
-geopandas==0.10.2
-kaleido==0.2.1
-geojson==2.5.0
-plotly==4.14.3
-h3==3.7.0
-pyyaml==5.4.1
-rtree==0.9.7
-openpyxl==3.0.9
+setuptools>=65.6.3
+geopandas>=0.10.2
+kaleido>=0.2.1
+geojson>=2.5.0
+plotly>=4.14.3
+h3>=3.7.0
+pyyaml>=5.4.1
+rtree>=0.9.7
+openpyxl>=3.0.9
+pandas>=1.5.3
+numpy>=1.24.1
+shapely>=2.0.0
+pyproj>=3.4.1
+fiona>=1.8.22
