+++ .github/workflows/geohexviz-automated-tests-conda.yml
@@ -14,7 +14,7 @@ jobs:
         fail-fast: false
         matrix:
             os: [ubuntu-latest, macos-latest, windows-latest]
-            python-version: [ '3.7', '3.8', '3.9' ]
+            python-version: [ '3.7', '3.8' ]
     steps:
       - uses: actions/checkout@v2
       - uses: conda-incubator/setup-miniconda@v2
