+++ .github/workflows/geohexviz-automated-tests-pip.yml
@@ -14,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: ['3.7', '3.8', '3.9']
+        python-version: ['3.7', '3.8']
     name: Test Build ${{ matrix.os }}, Python ${{ matrix.python-version }}
     steps:
       - uses: actions/checkout@v2
