+++ requirements.txt
@@ -7,7 +7,7 @@ h3>=3.7.0
 pyyaml>=5.4.1
 rtree>=0.9.7
 openpyxl>=3.0.9
-pandas>=1.5.3
+pandas==1.5.3
 numpy>=1.24.1
 shapely>=2.0.0
 pyproj>=3.4.1
