### 2023-01-26 f7c581e
Trying another speedup.
