+++ .github/workflows/geohexviz-automated-tests-conda.yml
@@ -1,4 +1,5 @@
 name: CI Tests (Conda)
+
 on:
   push:
     branches:
@@ -7,6 +8,10 @@ on:
   pull_request:
     branches:
       - master
+
+env:
+  CACHE_NUMBER: 0  # increase to reset cache manually
+
 jobs:
   miniconda:
     name: ${{ matrix.os }}, Python ${{ matrix.python-version }}
@@ -32,33 +37,33 @@ jobs:
 
         uses: conda-incubator/setup-miniconda@v2
         with:
-          installer-url: https://github.com/conda-forge/miniforge/releases/download/4.9.2-2/Mambaforge-Linux-x86_64.sh
-          channel-priority: true
-          use-only-tar-bz2: true
-          environment-file: environment.yml
-          activate-environment: test
-          python-version: ${{ matrix.python-version }}
-          auto-activate-base: false
+          miniforge-variant: Mambaforge
+          miniforge-version: latest
+          active-environment: my-env
+          use-mamba: true
 
-      - name: Anaconda Login
-        run: |
-          # Bypass the installation of anaconda-client
-          TOKEN_DIR="$HOME/.config/binstar"
-          TOKEN_PATH="$TOKEN_DIR/https%3A%2F%2Fapi.anaconda.org.token"
-          mkdir -p $TOKEN_DIR
-          echo -e "${ANACONDA_USER_TOKEN}\c" > $TOKEN_PATH
+      - name: Set cache date and cache
+        run: echo "DATE=$(date +'%Y%m%d')" >> $GITHUB_ENV
 
-      - name: Install Dependencies
-        run: mamba env update -f env.yml
+      - uses: actions/cache@v2
+        with:
+          path: ${{ matrix.prefix }}
+          key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}-${{ env.DATE }}-${{ env.CACHE_NUMBER }}
+        id: cache
 
+      - name: Update environment
+        run: mamba env update -n my-env -f environment.yml
+        if: steps.cache.outputs.cache-hit != 'true'
+
+      - name: Display conda info
       - shell: bash -l {0}
         run: |
-          conda info
-          conda list
+          mamba info
+          mamba list
       - name: Lint with flake8
         shell: bash -l {0}
         run: |
-            conda install flake8
+            mamba install flake8
             python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
             python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
       - name: Run tests with unittest
