+++ .github/workflows/geohexviz-automated-tests-conda.yml
@@ -45,7 +45,7 @@ jobs:
       - name: Set cache date and cache
         run: echo "DATE=$(date +'%Y%m%d')" >> $GITHUB_ENV
 
-      - uses: actions/cache@v2
+        uses: actions/cache@v2
         with:
           path: ${{ matrix.prefix }}
           key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}-${{ env.DATE }}-${{ env.CACHE_NUMBER }}
