+++ .github/workflows/geohexviz-automated-tests-conda.yml
@@ -22,8 +22,7 @@ jobs:
             os: [ubuntu-latest, macos-latest, windows-latest]
             python-version: [ '3.7', '3.8' ]
     steps:
-      - name: Checkout the code
-        uses: actions/checkout@v2
+      - uses: actions/checkout@v2
 
       - name: Cache Conda
         uses: actions/cache@v1
@@ -60,6 +59,7 @@ jobs:
         run: |
           mamba info
           mamba list
+
       - name: Lint with flake8
         shell: bash -l {0}
         run: |
