+++ .github/workflows/geohexviz-automated-tests-conda.yml
@@ -21,19 +21,11 @@ jobs:
         matrix:
             os: [ubuntu-latest, macos-latest, windows-latest]
             python-version: [ '3.7', '3.8' ]
+
     steps:
       - uses: actions/checkout@v2
 
-      - name: Cache Conda
-        uses: actions/cache@v1
-        env:
-          CACHE_NUMBER: 2
-        with:
-          path: ~/conda_pkgs_dir
-          key: conda-${{ runner.os }}-${{ env.CACHE_NUMBER }}-${{ hashFiles('env.yml') }}
-
       - name: Setup Conda
-
         uses: conda-incubator/setup-miniconda@v2
         with:
           miniforge-variant: Mambaforge
@@ -67,6 +59,7 @@ jobs:
             mamba install flake8
             python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
             python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+
       - name: Run tests with unittest
         shell: bash -l {0}
         run: |
