+++ .github/workflows/tester.yml
@@ -59,6 +59,13 @@ jobs:
         run: mamba env update -n my-env -f environment.yml
         if: steps.cache.outputs.cache-hit != 'true'
 
+      - name: Lint the code
+        shell: bash -l {0}
+        run: |
+          mamba install flake8
+          python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+          python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+
       - name: Run tests
         shell: bash -l {0}
-        run: pytest ./tests
+        run: python -m unittest discover
