+++ .github/workflows/geohexviz-automated-tests-conda.yml
@@ -56,9 +56,9 @@ jobs:
       - name: Lint with flake8
         shell: bash -l {0}
         run: |
-            mamba install flake8
-            python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-            python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
+          mamba install flake8
+          python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
+          python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
 
       - name: Run tests with unittest
         shell: bash -l {0}
