+++ .github/workflows/tester.yml
@@ -1,4 +1,4 @@
-name: ci
+name: CI-Tests (Conda)
 
 on:
   push:
@@ -17,24 +17,13 @@ env:
 
 jobs:
   build:
-
-    strategy:
-      matrix:
-        include:
-          - os: ubuntu-latest
-            label: linux-64
-            prefix: /usr/share/miniconda3/envs/my-env
-
-          - os: macos-latest
-            label: osx-64
-            prefix: /Users/runner/miniconda3/envs/my-env
-
-          - os: windows-latest
-            label: win-64
-            prefix: C:\Miniconda3\envs\my-env
-
-    name: ${{ matrix.label }}
+    name: ${{ matrix.os }}, Python ${{ matrix.python-version }}
     runs-on: ${{ matrix.os }}
+    strategy:
+        fail-fast: false
+        matrix:
+            os: [ubuntu-latest, macos-latest, windows-latest]
+            python-version: [ '3.7', '3.8' ]
     steps:
       - uses: actions/checkout@v2
 
@@ -55,11 +44,11 @@ jobs:
           key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}-${{ env.DATE }}-${{ env.CACHE_NUMBER }}
         id: cache
 
-      - name: Update environment
+      - name: Update mamba environment
         run: mamba env update -n my-env -f environment.yml
         if: steps.cache.outputs.cache-hit != 'true'
 
-      - name: Lint the code
+      - name: Lint with flake8
         shell: bash -l {0}
         run: |
           mamba install flake8
