+++ .github/workflows/tester.yml
@@ -20,8 +20,8 @@ jobs:
 
     strategy:
       matrix:
-        python-version: [ '3.7', '3.8' ]
         include:
+          - python-version: [ '3.7', '3.8' ]
           - os: ubuntu-latest
             label: linux-64
             prefix: /usr/share/miniconda3/envs/my-env
@@ -32,7 +32,6 @@ jobs:
 
           - os: windows-latest
             label: win-64
-
             prefix: C:\Miniconda3\envs\my-env
 
     name: ${{ matrix.label }}
