+++ .github/workflows/tester.yml
@@ -13,15 +13,15 @@ on:
     - cron: '0 0 * * *'
 
 env:
-  CACHE_NUMBER: 0  # increase to reset cache manually
+  CACHE_NUMBER: 1  # increase to reset cache manually
 
 jobs:
   build:
 
     strategy:
       matrix:
+        python-version: [ '3.7', '3.8' ]
         include:
-          - python-version: [ '3.7', '3.8' ]
           - os: ubuntu-latest
             label: linux-64
             prefix: /usr/share/miniconda3/envs/my-env
