+++ .github/workflows/tester.yml
@@ -20,8 +20,7 @@ jobs:
 
     strategy:
       matrix:
-        name: ${{ matrix.label }}, Python ${{ matrix.python-version }}
-        runs-on: ${{ matrix.os }}
+
         python-version: [ '3.7', '3.8' ]
         include:
           - os: ubuntu-latest
@@ -35,6 +34,9 @@ jobs:
           - os: windows-latest
             label: win-64
             prefix: C:\Miniconda3\envs\my-env
+
+    name: ${{ matrix.label }}, Python ${{ matrix.python-version }}
+    runs-on: ${{ matrix.os }}
 
     steps:
       - uses: actions/checkout@v2
