+++ .github/workflows/tester.yml
@@ -1,91 +1,0 @@
-name: ci
-
-on:
-  push:
-    branches:
-      - '*'
-
-  pull_request:
-    branches:
-      - '*'
-
-  schedule:
-    - cron: '0 0 * * *'
-
-env:
-  CACHE_NUMBER: 1  # increase to reset cache manually
-
-jobs:
-  build:
-
-    strategy:
-      matrix:
-        include:
-          - os: ubuntu-latest
-            label: linux-64
-            python-version: '3.7'
-            prefix: /usr/share/miniconda3/envs/my-env
-
-          - os: ubuntu-latest
-            label: linux-64
-            python-version: '3.8'
-            prefix: /usr/share/miniconda3/envs/my-env
-
-          - os: macos-latest
-            label: osx-64
-            python-version: '3.7'
-            prefix: /Users/runner/miniconda3/envs/my-env
-
-          - os: macos-latest
-            label: osx-64
-            python-version: '3.8'
-            prefix: /Users/runner/miniconda3/envs/my-env
-
-          - os: windows-latest
-            label: win-64
-            python-version: '3.7'
-            prefix: C:\Miniconda3\envs\my-env
-
-          - os: windows-latest
-            label: win-64
-            python-version: '3.8'
-            prefix: C:\Miniconda3\envs\my-env
-
-    name: ${{ matrix.label }}, Python ${{ matrix.python-version }}
-    runs-on: ${{ matrix.os }}
-
-    steps:
-      - uses: actions/checkout@v2
-
-      - name: Setup Mambaforge
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          python-version: ${{ matrix.python-version }}
-          miniforge-variant: Mambaforge
-          miniforge-version: latest
-          activate-environment: my-env
-          use-mamba: true
-
-      - name: Set cache date
-        run: echo "DATE=$(date +'%Y%m%d')" >> $GITHUB_ENV
-
-      - uses: actions/cache@v2
-        with:
-          path: ${{ matrix.prefix }}
-          key: ${{ matrix.label }}-conda-${{ hashFiles('environment.yml') }}-${{ env.DATE }}-${{ env.CACHE_NUMBER }}
-        id: cache
-
-      - name: Update environment
-        run: mamba env update -n my-env -f environment.yml
-        if: steps.cache.outputs.cache-hit != 'true'
-
-      - name: Lint the code
-        shell: bash -l {0}
-        run: |
-          mamba install flake8
-          python -m flake8 . --count --select=E9,F63,F7,F82 --show-source --statistics
-          python -m flake8 . --count --exit-zero --max-complexity=10 --max-line-length=127 --statistics
-
-      - name: Run tests
-        shell: bash -l {0}
-        run: python -m unittest discover
