+++ requirements.txt
@@ -9,6 +9,6 @@ rtree>=0.9.7
 openpyxl>=3.0.9
 pandas
 numpy>=1.24.1
-shapely>=2.0.0
+shapely
 pyproj>=3.4.1
 fiona>=1.8.22
