+++ requirements.txt
@@ -8,7 +8,7 @@ pyyaml>=5.4.1
 rtree>=0.9.7
 openpyxl>=3.0.9
 pandas
-numpy>=1.24.1
+numpy
 shapely
 pyproj>=3.4.1
 fiona>=1.8.22
