+++ requirements.txt
@@ -10,5 +10,5 @@ openpyxl>=3.0.9
 pandas
 numpy
 shapely
-pyproj>=3.4.1
-fiona>=1.8.22
+pyproj
+fiona
