### 2023-01-26 def1eeb
Revert "Create main.yml"
This reverts commit 6ba8ea8eda8033bf2a84775e4e49d6ab20558112.

# Conflicts:
#	.github/workflows/sphinx-docs.yml
