+++ .github/workflows/sphinx-docs.yml
@@ -1,43 +1,0 @@
-name: Sphinx docs to gh-pages
-
-on:
-  push:
-    branches:
-      - main
-
-jobs:
-  sphinx_docs_generate:
-    runs-on: ubuntu-latest
-    steps:
-      - uses: actions/checkout@v1
-      - uses: ammaraskar/sphinx-action@master
-        with:
-          docs-folder: "docs/sphinx"
-
-  sphinx_docs_to_gh-pages:
-    runs-on: ubuntu-latest
-    name: Sphinx docs to gh-pages
-    steps:
-      - uses: actions/checkout@v2
-        with:
-          fetch-depth: 0
-      - name: Make conda environment
-        uses: conda-incubator/setup-miniconda@v2
-        with:
-          python-version: 3.7
-          environment-file: devtools/conda-envs/docs_env.yaml
-          auto-update-conda: false
-          auto-activate-base: false
-          show-channel-urls: true
-      - name: Installing the library
-        shell: bash -l {0}
-        run: |
-          python setup.py install
-      - name: Running the Sphinx to gh-pages Action
-        uses: uibcdf/action-sphinx-docs-to-gh-pages@v1.0.0
-        with:
-          branch: main
-          dir_docs: docs
-          sphinxapiopts: '--separate -o . ../'
-          sphinxapiexclude: '../*setup* ../*.ipynb'
-          sphinxopts: ''
