+++ setup.py
@@ -10,8 +10,8 @@ setup(
     long_description=readme,
     long_description_content_type="text/x-rst",
     url="",
-    author="Tony Marco Abou Zeidan",
-    author_email="tony.azp25@gmail.com",
+    author="Tony Marco Abou Zeidan, Mark Rempel",
+    author_email="tony.azp25@gmail.com, mark.rempel@forces.gc.ca",
     license="MIT",
     classifiers=[
         "Programming Language :: Python :: 3 :: Only",
@@ -23,18 +23,19 @@ setup(
         "Topic :: Scientific/Engineering"
     ],
     keywords='visualization,geospatial,hexagonal binning',
-    python_requires=[">=3.7", "<=3.8.16"],
+    python_requires=">=3.7, <=3.8.16",
     packages=find_packages(exclude=["docs", "tests"]),
     include_package_data=True,
     install_requires=[
         "h3==3.7.0",
-        "shapely==^1.7.1",
+        "shapely==1.8.5",
         "geojson==2.5.0",
         "plotly==4.14.3",
         "kaleido==0.2.1",
         "pyyaml==5.4.1",
         "rtree==0.9.7",
-        "openpyxl==3.0.9"   # required for XLSX input
+        "openpyxl==3.0.9",  # required for XLSX input
+        "pdfcropmargins"
     ],
     extras_require={
         "dev": [
