+++ .github/workflows/geohexviz-automated-tests-pip.yml
@@ -25,8 +25,8 @@ jobs:
         uses: actions/setup-python@v2
         with:
           python-version: ${{ matrix.python-version }}
-          cache: 'CI-pip-' ${{ matrix.os }}
-
+          cache: 'pip'
+          cache-dependency-path: '**/requirements*.txt'
       - name: Install dependencies
         run: |
           python -m pip install --upgrade pip
