### 2023-01-28 88515a9
Address issue #1 and #2
