+++ setup.py
@@ -17,7 +17,7 @@ setup(
         "Programming Language :: Python :: 3 :: Only",
         "Programming Language :: Python :: 3",
         "License :: OSI Approved :: MIT License",   
-        "Operating System :: OS Independent",  # TODO: ensure deployment on other OS
+        "Operating System :: OS Independent",  
         "Topic :: Scientific/Engineering :: Information Analysis",
         "Topic :: Scientific/Engineering :: Visualization",
         "Topic :: Scientific/Engineering"
@@ -27,24 +27,20 @@ setup(
     packages=find_packages(exclude=["docs", "tests"]),
     include_package_data=True,
     install_requires=[
-        "h3==3.7.0",
-        "shapely==1.8.5",
-        "geojson==2.5.0",
-        "plotly==4.14.3",
-        "kaleido==0.2.1",
-        "pyyaml==5.4.1",
-        "rtree==0.9.7",
-        "openpyxl==3.0.9",  # required for XLSX input
+        "h3>=3.7.0",
+        "shapely>=1.8.5",
+        "geojson>=2.5.0",
+        "plotly>=4.14.3",
+        "kaleido>=0.2.1",
+        "pyyaml>=5.4.1",
+        "rtree>=0.9.7",
+        "openpyxl>=3.0.9",
         "pdfcropmargins"
     ],
     extras_require={
         "dev": [
             "pip",
-            "Sphinx",
-            "pdfcropmargins==1.0.5"
-        ],
-        "pdf-crop": [
-            "pdfcropmargins~=1.0.5"
+            "Sphinx"
         ]
     },
     entry_points={
