+++ environment.yml
@@ -4,17 +4,17 @@ channels:
   - defaults
   - plotly
 dependencies:
-  - python<=3.8.16
+  - python<=3.8.16,>=3.7
   - pip
   - geopandas
   - setuptools
-  - python-kaleido
-  - geojson
-  - plotly
-  - h3-py
-  - pyyaml
-  - rtree
-  - openpyxl
-  - shapely
+  - python-kaleido>=0.2.1
+  - geojson>=2.5.0
+  - plotly>=4.14.3
+  - h3-py>=3.7.0
+  - pyyaml>=5.4.1
+  - rtree>=0.9.7
+  - openpyxl>=3.0.9
+  - shapely>=1.8.5
   - pip:
     - pdfcropmargins
