### 2023-01-28 mrempel/geohexviz#2 @mrempel
We have updated the environment.yml and setup.py files so that the version numbers of the required packages agree between the files. For example, both now state that shapely>=1.8.5. In addition, as appropriate the packages requirements have been modified to be >= a version number as oppose to being a specific version number.
