### 2023-01-28 mrempel/geohexviz#3 @mrempel
We have modified the examples such that the file placeholders have been replaced with the actual files to be called to execute the examples. For example, when a user executes `python python_walkthrough.py` in the forest_fire example directory, the corresponding visualization will be created as expected. This has also be done for the .md files and notebooks resident in the example directories.
