+++ pyproject.toml
@@ -13,7 +13,7 @@ readme = "README.rst"
 requires-python = ">=3.7, <=3.8.16"
 keywords = ["visualization","geospatial","hexagonal binning"]
 license = {text = "MIT"}
-dynamic = ['version']
+version="1.0.0"
 classifiers=[
     "Programming Language :: Python :: 3 :: Only",
     "Programming Language :: Python :: 3",
