### 2023-02-02 #5073 @kaustavbhattacharjee
@martinfleis I have completed my review, and I think this is ready for publication within JOSS.
I was able to run the program quite easily. Here are a few minor comments for the authors @mrempel @tony-zeidan :

- I had some doubts about the contributions of @mrempel. But it was clarified through the previous conversation on this thread.
- The image bombings.pdf is not rendered properly in [paper.md](https://github.com/mrempel/geohexviz/blob/master/docs/paper/paper.md). I would recommend either converting it to an image or embedding PDF using HTML code.
- I was also wondering if this package can be published to PyPI so that users can install it directly through pip
