### 2023-02-03 f7c0c2b
Update paper.md to use png
