### 2023-02-03 #5073 @martinfleis
> there are some issues for us that come along with publishing the package to PyPI

Can you elaborate? We may be able to help here. I would very much like to see the package published there before the acceptance, even though it is not a formal requirement of the JOSS.
