### 2023-02-03 7d38e0e
Update README.rst to support PyPI
