### 2023-02-03 c2c07b3
Update README.rst for PyPI install
