### 2023-02-03 #5073 @mrempel
@martinfleis @kaustavbhattacharjee - Regarding the two latter comments from @kaustavbhattacharjee:

- Figure in paper.md: The figure has been changed from a PDF to PNG, and thus now renders properly (both in the a browser and in the compiled PDF). This has been pushed to the master branch.

- Publish to PyPI: The difficulty that Tony was referring to was that geopandas is more difficult to install from pip, and our install process uses the conda-forge channel. That said, we have now pushed GeoHexViz to PyPI (see https://pypi.org/project/geohexviz/) and the install instructions have been modified to provide two options - install from PyPI (option 1), and install from GitHub (option 2). Option 1 requires the user to install geopandas separately, and instructions are provided in README.rst to acomplish this task.

Please let us know if there are any further questions or comments regarding our submission.
