### 2023-02-03 #5073 @martinfleis
As a geopandas maintainer I think I know what you're talking about :). Thanks for pushing it to PyPI! Do you also want to make it available on conda-forge? Happy to help setting up the recipe.
