### 2023-02-03 #5073 @mrempel
@tony-zeidan and I were talking about that - it is an enhancement listed in our issues - but didn't want to delay the final decision regarding publication. If it possible to make it available on conda-forge without too much work, then we would be happy to do it and any help setting up a recipe would be appreciated.
