### 2023-02-04 #5073 @martinfleis
@mrempel I'll have a look and let you know, it is not that much work. We don't have to wait with the acceptance for it though :).

Thank you @gassmoeller and @kaustavbhattacharjee for your reviews!

The submission is now almost ready to be published.

@mrempel the next steps you need to do now:

- [ ] Make a new tagged release (or let me know if the latest one contains all the changes) of the software, and list the version tag of the archived version here.
- [ ]  Check that the archival deposit (e.g., in Zenodo) has the correct metadata. In particular, the title and author list should be identical to those of the paper.
