+++ environment.yml
@@ -4,7 +4,7 @@ channels:
   - defaults
   - plotly
 dependencies:
-  - python<=3.8.16,>=3.7
+  - python>=3.7
   - pip
   - geopandas
   - setuptools
