+++ .github/workflows/geohexviz-automated-tests-pip.yml
@@ -2,11 +2,10 @@ name: CI Tests (Pip)
 on:
   push:
     branches:
-      - master
-      - buffer
+      - '*'
   pull_request:
     branches:
-      - master
+      - '*'
 
 jobs:
   build:
@@ -15,7 +14,7 @@ jobs:
       fail-fast: false
       matrix:
         os: [ubuntu-latest, macos-latest]
-        python-version: ['3.7', '3.8']
+        python-version: ['3.7', '3.8', '3.9', '3.10', '3.11']
     name: Test Build ${{ matrix.os }}, Python ${{ matrix.python-version }}
     steps:
       - name: Checkout code
