+++ pyproject.toml
@@ -13,7 +13,7 @@ readme = "README.rst"
 requires-python = ">=3.7"
 keywords = ["visualization","geospatial","hexagonal binning"]
 license = {text = "MIT"}
-version="1.0.0"
+version="1.0.1"
 classifiers=[
     "Programming Language :: Python :: 3",
     "Programming Language :: Python :: 3.7",
