### 2023-02-06 b4a81fb
Link geohexviz version to pypi
