### 2023-02-07 #5073 @editorialbot
:wave: @openjournals/sbcs-eics, this paper is ready to be accepted and published.
