### 2023-02-15 #5073 @mrempel
@martinfleis @arfon  Good day - It has been brought to our attention that there is a typo in the title of the article - it reads "GeoHexViz: A Python package for the visualizing hexagonally binned geospatial data", whereas it should read "GeoHexViz: A Python package for visualizing hexagonally binned geospatial data". The "the" should not be in the title.

I went back through the archive of the geohexviz repository on github, and this error was in the version of the paper that was originally submitted. Thus, this is our mistake.

Is there a way that this could be fixed? I can adjust the paper.md file in the geohexviz repository, but can the published version be updated?
